test(views): rename file and add /update_repository files
